synq: extract transport-agnostic LSP dispatcher
The LSP request handlers were already pure functions of the request and
the LspHost, but they were locked inside the stdio server: the lifecycle
lived in lsp_server::Connection and publishDiagnostics was pushed
straight into the stdio sender. That made it impossible to run the
server over any other transport, such as a web worker channel.

Move the handlers into LspDispatcher, which owns the lifecycle
(initialize/shutdown/exit) and returns outgoing messages as values —
responses plus server-initiated notifications. LspServer is now a thin
stdio loop over the dispatcher; handle_json accepts and produces plain
JSON strings for transports without typed messages.

Groundwork for an LSP entry point in syntaqlite-wasm (#298).
